Move protocols to esmerald.core
