+++ docs_src/middleware/sample.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict
 
-from esmerald.protocols.middleware import MiddlewareProtocol
+from esmerald.core.protocols.middleware import MiddlewareProtocol
 from esmerald.types import ASGIApp, Receive, Scope, Send
