+++ esmerald/conf/global_settings.py
@@ -20,11 +20,11 @@ from esmerald.core.config import (
 from esmerald.core.config.asyncexit import AsyncExitConfig
 from esmerald.core.datastructures import Secret
 from esmerald.core.interceptors.types import Interceptor
+from esmerald.core.protocols.cache import CacheBackend
 from esmerald.encoders import Encoder
 from esmerald.openapi.schemas.v3_1_0 import Contact, License, SecurityScheme
 from esmerald.permissions.types import Permission
 from esmerald.pluggables import Extension, Pluggable
-from esmerald.protocols.cache import CacheBackend
 from esmerald.routing import gateways
 from esmerald.types import (
     APIGateHandler,
