+++ esmerald/logging.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any, cast
 
-from esmerald.protocols.logging import LoggerProtocol
+from esmerald.core.protocols.logging import LoggerProtocol
 
 
 class LoggerProxy:
