+++ tests/dependencies/test_factory.py
@@ -3,7 +3,7 @@ from typing import Any
 import pytest
 
 from esmerald import Factory, Gateway, Inject, JSONResponse, get
-from esmerald.protocols.asyncdao import AsyncDAOProtocol
+from esmerald.core.protocols.asyncdao import AsyncDAOProtocol
 from esmerald.testclient import create_client
